requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# CrewAI Multi-Agent System (Latest Available)
crewai

# HTTP Requests & Environment
requests
python-dotenv

# Data Processing (for future integrations)
pandas
openpyxl

# Airtable Integration
pyairtable

# Date/Time Handling
python-dateutil